refactor: extract ArgumentDialog to own file
